+++ docs/changes.rst
@@ -9,7 +9,7 @@ To be released.
 - Added :class:`sider.types.Date` type.
 - Added :class:`sider.datetime.FixedOffset` tzinfo subtype.
 - Added :class:`sider.types.DateTime` and
-  :class:`~sider.types.TZDatetime` types.
+  :class:`~sider.types.TZDateTime` types.
 
 
 Version 0.1.0